Refactored CSS file to have separate CSS subfiles to make debugging easier
