--- client/src/App.css
@@ -1,421 +1,0 @@
-:root{
-    --accent:#A6192E;
-    --bg:#111315;
-    --border:#595151;
-    --muted:#5E5454;
-    --panel:#2D2828;
-    --panelLight:#403939;
-    --text:#E6DFDF;
-    --white:#FFFFFF;
-}
-
-*{
-    box-sizing:border-box;
-}
-body, #root, .page {
-    background:var(--bg);
-    color:var(--text);
-    font-family:system-ui,-apple-system,Segoe UI,Roboto,Helvetica,Arial;
-    height:100%;
-    margin:0;
-}
-
-/* Top slim bar */
-.loginBar{
-    background:#2D2828;
-    border-bottom:1px solid var(--border);
-    display:flex;
-    justify-content:flex-end;
-    align-items:center;
-    padding:6px 16px;
-}
-.textBtn{
-    background:transparent;
-    border:0;
-    color:#c6cbd1;
-    font-weight: bold;
-    padding:6px 10px;
-    border-radius:999px;
-}
-.textBtn:disabled{
-    opacity:.9;
-}
-
-/* Header row */
-.header{
-    display:grid;
-    grid-template-columns: 180px 1fr auto;
-    gap:12px;
-    align-items:center;
-    padding:16px;
-    border-bottom:1px solid var(--border);
-}
-.logoPlaceholder{
-    height:40px;
-    width:172px;
-    border-radius:10px;
-    color:#595151;
-    background: #2D2828 url(images/mainLogo.png);
-    background-size: cover;
-}
-
-.searchWrap{
-    position:relative;
-}
-.search{
-    width:100%;
-    padding:12px 40px 12px 12px;
-    border-radius:12px;
-    background:#2D2828;
-    border:1px solid var(--border);
-    color:var(--text);
-}
-.search:disabled{
-    opacity:.9;
-}
-.searchIcon{
-    position:absolute;
-    right:10px;
-    top:50%;
-    transform:translateY(-50%);
-    opacity:.7;
-}
-
-.createBtn{
-    background:var(--accent);
-    color:var(--white);
-    border:0;
-    padding:12px 16px;
-    border-radius:16px;
-    font-weight:600;
-}
-.createBtn:disabled{
-    opacity:.9;
-}
-
-/* Main content layout */
-.content{
-    display:grid;
-    grid-template-columns: 400px 1fr;
-    gap:24px;
-    padding:24px 40px 40px;
-    margin:0;
-    width:100%;
-}
-
-/* Left column cards */
-.leftCol{
-    display:flex;
-    flex-direction:column;
-    gap:24px;
-}
-
-.card{
-    border-radius:16px;
-    overflow:hidden;
-    border:1px solid var(--border);
-}
-
-/* Report Card */
-
-.reportCard{
-    background:#ffffff;
-    color:#111;
-    padding:20px;
-    min-height:220px;
-}
-.reportCard h2{
-    margin:0 0 8px;
-}
-.muted{
-    color:#666;
-}
-
-.reportField {
-    margin: 10px 0;
-    font-size: 0.95rem;
-    color: #333;
-}
-
-.reportField strong {
-    display: block;
-    font-weight: 700;
-    color: #111;
-}
-
-.reportField span {
-    display: block;
-    color: #555;
-    margin-top: 2px;
-}
-
-.reportDescription {
-    margin-top: 4px;
-    color: #555;
-    font-style: italic;
-}
-
-.filtersCard{
-    background:var(--panel);
-    padding:20px;
-}
-.filtersCard h3{
-    margin:0 0 12px;
-}
-.filtersCard label{
-    display:block;
-    margin:10px 0;
-    color:#d6d9dd;
-}
-.filtersCard input{
-    margin-right:8px;
-}
-
-/* Map placeholder rectangle */
-.mapPanel{
-    background:#595151;
-    border:1px solid #E6DFDF;
-    border-radius:16px;
-    min-height:520px;
-    color:#403939;
-    display:flex;
-    align-items:center;
-    justify-content:center;
-    font-weight:600;
-}
-
-/* Media query to adjust as window size changes*/
-@media (max-width: 980px){
-    .content{
-        grid-template-columns:1fr;
-    }
-}
-
-/* Report Popup */
-.modalCard{
-    /* switch to dark to match site */
-    background: var(--panel);
-    color: var(--text);
-    width: min(720px, 94vw);
-    border-radius: 16px;
-    box-shadow: 0 10px 40px rgba(0,0,0,.45);
-    border: 1px solid var(--border);
-}
-
-.modalHeader{
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 18px 20px;
-    border-bottom: 1px solid var(--border);
-}
-
-.modalHeader h2{
-    margin: 0;
-    font-size: 1.2rem;
-    letter-spacing: .2px;
-}
-
-.modalClose{
-    background: transparent;
-    border: 0;
-    font-size: 1.1rem;
-    cursor: pointer;
-    color: #bbb;
-}
-.modalClose:hover{ color:#fff; }
-
-.modalBody{
-    padding: 20px;
-    color: var(--text);
-}
-
-/* Grid for the form */
-.formGrid{
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 16px 18px; /* row col */
-}
-
-.full{
-    grid-column: 1 / -1;
-}
-
-/* Field block */
-.field{
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-}
-
-.field label{
-    font-size: .92rem;
-    color: #e9e3e3;
-    letter-spacing: .2px;
-}
-
-/* Section title inside form */
-.fieldsetTitle{
-    margin-top: 4px;
-    font-weight: 700;
-    color: var(--white);
-    opacity: .92;
-}
-
-/* Input */
-.input{
-    background: var(--panelLight);
-    color: var(--text);
-    border: 1px solid var(--border);
-    border-radius: 12px;
-    padding: 10px 12px;
-    font: inherit;
-    outline: none;
-    transition: border-color .15s ease, box-shadow .15s ease;
-}
-
-.input::placeholder{
-    color: #bfb8b8;
-    opacity: .85;
-}
-
-.input:focus{
-    border-color: var(--accent);
-    box-shadow: 0 0 0 3px rgba(166,25,46,.25);
-}
-
-/* Text area */
-.textarea{
-    resize: vertical;
-    min-height: 110px;
-}
-
-/* Time row with segmented AM/PM */
-.timeRow{
-    display: flex;
-    gap: 12px;
-    align-items: center;
-}
-
-/* AM/PM */
-.segmented{
-    display: inline-flex;
-    background: var(--panelLight);
-    border: 1px solid var(--border);
-    border-radius: 12px;
-    overflow: hidden;
-}
-
-.segmentedItem{
-    position: relative;
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    padding: 8px 14px;
-    cursor: pointer;
-    user-select: none;
-    font-size: .9rem;
-    color: #ddd;
-}
-
-.segmentedItem + .segmentedItem{
-    border-left: 1px solid var(--border);
-}
-
-.segmentedItem input{
-    /* hide native radio but keep it accessible */
-    position: absolute;
-    opacity: 0;
-    inset: 0;
-    cursor: pointer;
-}
-
-.segmentedItem input:checked + span{
-    background: var(--accent);
-    color: var(--white);
-    padding: 6px 12px;
-    border-radius: 8px;
-}
-
-/* Checkboxes row */
-.checksRow{
-    display: flex;
-    flex-wrap: wrap;
-    gap: 10px 14px;
-}
-
-.check{
-    display: inline-flex;
-    align-items: center;
-    gap: 8px;
-    padding: 8px 10px;
-    background: var(--panelLight);
-    border: 1px solid var(--border);
-    border-radius: 10px;
-}
-
-.check input{
-    accent-color: var(--accent);
-}
-
-.actions{
-    display: flex;
-    justify-content: flex-end;
-    gap: 10px;
-    margin-top: 6px;
-}
-
-/* Buttons */
-.btn{
-    border: 1px solid transparent;
-    border-radius: 12px;
-    padding: 10px 14px;
-    font-weight: 600;
-    cursor: pointer;
-}
-
-.btn.primary{
-    background: var(--accent);
-    color: var(--white);
-}
-.btn.primary:hover{
-    filter: brightness(1.05);
-}
-
-.btn.secondary{
-    background: transparent;
-    color: var(--text);
-    border-color: var(--border);
-}
-.btn.secondary:hover{
-    background: rgba(255,255,255,.06);
-}
-
-/* Narrow screens: stack into one column */
-@media (max-width: 720px){
-    .formGrid{
-        grid-template-columns: 1fr;
-    }
-    .timeRow{
-        flex-direction: column;
-        align-items: stretch;
-    }
-    .segmented{
-        width: fit-content;
-    }
-}
-
-/* Centering position */
-.modalBackdrop{
-    position: fixed;
-    inset: 0;
-    background: rgba(0,0,0,.55);
-    display: grid;
-    place-items: center;
-    z-index: 1000;
-}
-
-.modalCard{
-    max-height: 90vh;
-    overflow: auto;
-}
