+++ client/src/Styles/App.css
@@ -1,0 +1,281 @@
+:root{
+    --accent:#A6192E;
+    --bg:#111315;
+    --border:#595151;
+    --text:#E6DFDF;
+    --white:#FFFFFF;
+    --muted:#5E5454;
+    --panel:#2D2828;
+    --panelLight:#403939;
+}
+
+*{
+    box-sizing:border-box;
+}
+body, #root, .page {
+    background:var(--bg);
+    color:var(--text);
+    font-family:system-ui,-apple-system,Segoe UI,Roboto,Helvetica,Arial;
+    height:100%;
+    margin:0;
+}
+
+/* HEADER */
+.header{
+    display:flex;
+    justify-content: space-between;
+    align-items:center;
+    padding:12px;
+    border-bottom:1px solid var(--border);
+    background-color: #fff;
+}
+
+.loginBtn{
+    background:transparent;
+    border:0;
+    color:#c6cbd1;
+    font-weight: bold;
+    padding:6px 10px;
+    border-radius:999px;
+}
+.loginBtn:disabled{
+    opacity:.9;
+}
+
+.schoolLogo{
+    height:40px;
+    width:172px;
+    background: transparent url(../images/mainLogo.png);
+    background-size: cover;
+}
+/* END HEADER */
+
+/* NAVBAR */
+.navBar{
+    display:flex;
+    justify-content: space-between;
+    align-items:center;
+    padding:2px;
+    border-bottom:1px solid var(--border);
+    gap: 10px;
+}
+
+.filtersCard{
+    background: #111111 var(--panel);
+    padding:8px;
+    width:100%;
+}
+/*END NAVBAR */
+
+/* MAIN CONTENT */
+.content{
+    display:grid;
+    grid-template-columns: 370px 1fr;
+    gap:10px;
+    padding:24px 20px 40px;
+    margin:0;
+    width:100%;
+    background: url("../images/school.jpg");
+    background-size: cover;
+}
+
+.leftCol{
+    display:flex;
+    flex-direction:column;
+}
+/* Map in separate css file */
+/* END MAIN CONTENT */
+
+/* BUTTONS */
+.btn{
+    border: 1px solid transparent;
+    border-radius: 12px;
+    padding: 10px 14px;
+    font-weight: 600;
+    cursor: pointer;
+}
+.btn.primary{
+    background: var(--accent);
+    color: var(--white);
+}
+.btn.primary:hover{
+    filter: brightness(1.05);
+}
+.btn.secondary{
+    background: transparent;
+    color: var(--text);
+    border-color: var(--border);
+}
+.btn.secondary:hover{
+    background: rgba(255,255,255,.06);
+}
+
+.createBtn{
+    background:var(--accent);
+    color:var(--white);
+    padding:5px 16px;
+    border-radius:16px;
+    font-weight:600;
+}
+.createBtn:disabled{
+    opacity:.9;
+}
+/* END BUTTONS */
+
+/* FORM */
+.formGrid{
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    gap: 16px 18px; /* row col */
+}
+
+.full{
+    grid-column: 1 / -1;
+}
+
+/* Field block */
+.field{
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+}
+
+.field label{
+    font-size: .92rem;
+    color: #e9e3e3;
+    letter-spacing: .2px;
+}
+
+/* Section title inside form */
+.fieldsetTitle{
+    margin-top: 4px;
+    font-weight: 700;
+    color: var(--white);
+    opacity: .92;
+}
+
+/* Input */
+.input{
+    background: var(--panelLight);
+    color: var(--text);
+    border: 1px solid var(--border);
+    border-radius: 12px;
+    padding: 10px 12px;
+    font: inherit;
+    outline: none;
+    transition: border-color .15s ease, box-shadow .15s ease;
+}
+
+.input::placeholder{
+    color: #bfb8b8;
+    opacity: .85;
+}
+
+.input:focus{
+    border-color: var(--accent);
+    box-shadow: 0 0 0 3px rgba(166,25,46,.25);
+}
+
+/* Text area */
+.textarea{
+    resize: vertical;
+    min-height: 110px;
+}
+
+/* Time row with segmented AM/PM */
+.timeRow{
+    display: flex;
+    gap: 12px;
+    align-items: center;
+}
+
+/* AM/PM */
+.segmented{
+    display: inline-flex;
+    background: var(--panelLight);
+    border: 1px solid var(--border);
+    border-radius: 12px;
+    overflow: hidden;
+}
+
+.segmentedItem{
+    position: relative;
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    padding: 8px 14px;
+    cursor: pointer;
+    user-select: none;
+    font-size: .9rem;
+    color: #ddd;
+}
+
+.segmentedItem + .segmentedItem{
+    border-left: 1px solid var(--border);
+}
+
+.segmentedItem input{
+    /* hide native radio but keep it accessible */
+    position: absolute;
+    opacity: 0;
+    inset: 0;
+    cursor: pointer;
+}
+
+.segmentedItem input:checked + span{
+    background: var(--accent);
+    color: var(--white);
+    padding: 6px 12px;
+    border-radius: 8px;
+}
+
+/* Checkboxes row */
+.checksRow{
+    display: flex;
+    flex-wrap: wrap;
+    gap: 10px 14px;
+}
+
+.check{
+    display: inline-flex;
+    align-items: center;
+    gap: 8px;
+    padding: 8px 10px;
+    background: var(--panelLight);
+    border: 1px solid var(--border);
+    border-radius: 10px;
+}
+
+.check input{
+    accent-color: var(--accent);
+}
+
+.actions{
+    display: flex;
+    justify-content: flex-end;
+    gap: 10px;
+    margin-top: 6px;
+}
+/* END FORM */
+
+/* MEDIA QUERY */
+@media (max-width: 720px){
+    .formGrid{
+        grid-template-columns: 1fr;
+    }
+    .timeRow{
+        flex-direction: column;
+        align-items: stretch;
+    }
+    .segmented{
+        width: fit-content;
+    }
+}
+
+/* Media query to adjust as window size changes*/
+@media (max-width: 980px){
+    .content{
+        grid-template-columns:1fr;
+    }
+}
+/* END MEDIA QUERY */
