+++ client/src/Styles/ReportCard.css
@@ -1,0 +1,32 @@
+/* Report Card */
+.reportCard{
+    background:#ffffff;
+    color:#111;
+    padding:20px;
+    min-height:220px;
+    border-radius: 16px;
+}
+.reportCard h2{
+    margin:0 0 8px;
+}
+
+.reportField {
+    margin: 10px 0;
+    font-size: 0.95rem;
+    color: #333;
+}
+.reportField strong {
+    display: block;
+    font-weight: 700;
+    color: #111;
+}
+.reportField span {
+    display: block;
+    color: #555;
+    margin-top: 2px;
+}
+.reportDescription {
+    margin-top: 4px;
+    color: #555;
+    font-style: italic;
+}
